healthcare-app/requirement.txt
# Core Requirements
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
python-dotenv==1.0.0

# Database & ORM
SQLAlchemy==2.0.15
psycopg2-binary==2.9.6  # For PostgreSQL
pymysql==1.0.3  # For MySQL/MariaDB

# Authentication & Security
PyJWT==2.7.0
bcrypt==4.0.1
Flask-Limiter==2.8.1

# API Documentation
flask-restx==1.1.0
apispec==6.3.0

# Email & Notifications
Flask-Mail==0.9.1
celery==5.2.7

# Caching
redis==4.5.5
Flask-Caching==2.0.2

# Data Validation
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1

# Utility Libraries
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Development & Testing (not for production)
pytest==7.3.2
pytest-cov==4.1.0
Faker==18.10.1
flake8==6.0.0
black==23.7.0